+++ src/fbe/mainfrm.h
@@ -16,6 +16,7 @@
 #include "document\\recent\\RecentDocumentsController.h"
 #include "scripts\\ScriptUiController.h"
 #include "plugins\\PluginUiController.h"
+#include "plugins\\PluginExecutionController.h"
 
 #include "atlctrlsext.h"
 
@@ -237,6 +238,7 @@ public:
 
 	// Plugins support
 	PluginUiController m_plugins;
+	PluginExecutionController m_plugin_execution;
 	void InitializeExtensionUi();
 	void InitializeScripts();
 	void InitializeBundledPlugins();